% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adm2pmml.R
\name{adm2pmml}
\alias{adm2pmml}
\title{Generates PMML files from ADM models.}
\usage{
adm2pmml(
  datamart,
  destDir = ".",
  tmpDir = NULL,
  verbose = (!is.null(tmpDir)),
  explanations = reasonCodesAboveMin(3)
)
}
\arguments{
\item{datamart}{Model and predictor data from the ADM datamart.}

\item{destDir}{Directory to create the PMML files in. Defaults to the current
directory.}

\item{tmpDir}{Directory for temporary files. Defaults to \code{NULL}
indicating no temp files will be left behind.}

\item{verbose}{Whether to print the steps it's performing. Defaults to
\code{TRUE} if the \code{tmpDir} is set.}

\item{explanations}{Data structure that represents the type and number
of explanations as returned by the \code{reasonCodes...} functions.}
}
\value{
Creates the PMML file(s) in the given destination and returns a list
  of model ID's per file.
}
\description{
Every model configuration becomes one big PMML file. The big ensemble contains
score cards for every model partition. The ensemble switches between the
models based on context key values.
}
\examples{
\dontrun{
  dm <- ADMDatamart("~/Downloads")

  adm2pmml(dm, verbose=TRUE)

  adm2pmml(dm, verbose=TRUE, explanations = reasonCodesAboveMean(4))

}
}
