% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{adm_datamart}
\alias{adm_datamart}
\title{Sample ADM Data Mart model and predictor data.}
\format{
A \code{list} with two \code{data.table} objects.
\describe{
  \item{modeldata}
  \item{predictordata}
}
}
\source{
A sample Pega application
}
\usage{
adm_datamart
}
\description{
Standard \code{list} datastructure as produced by \code{ADMDatamart}.
}
\keyword{datasets}
