% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{admdatamart_binning}
\alias{admdatamart_binning}
\title{Sample predictor binning snapshots from the ADM Data Mart.}
\format{
A \code{data.table} with 1755 rows and 32 variables:
\describe{
  \item{ModelID}{Model ID}
  \item{SnapshotTime}{Snapshot time}
  \item{PredictorName}{Name of the predictor}
  ...
}
}
\source{
A sample Pega application
}
\usage{
admdatamart_binning
}
\description{
A \code{data.table} containing a snapshot of the predictor binning of the ADM models in \code{admdatamart_models}.
}
\keyword{datasets}
