% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportUtilsJSON.R
\name{admJSONFactoryToBinning}
\alias{admJSONFactoryToBinning}
\title{Turns an ADM Factory JSON into an easy-to-process binning table.}
\usage{
admJSONFactoryToBinning(factoryJSON, modelname = "Dummy")
}
\arguments{
\item{factoryJSON}{JSON string for a single model, usually the pyfactory field from the table returned by getModelsFromJSONTable.}

\item{modelname}{Optional name for the model, filled in into the modelid field of the binning table.}
}
\value{
A data.table containing the flattened-out predictor binning.
}
\description{
It turns the ADM Factory JSON in the exact same format as the datamart binning table which includes aggregating
from the internal sub-bin model (to support individual field values) back to aggregate values.
}
\examples{
\dontrun{admFactory <- getModelsFromJSONTable(conn)
binning <- admJSONFactoryToBinning(admFactory$pyfactory[1])}
}
