% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adm.R
\name{admVarImp}
\alias{admVarImp}
\title{Calculate variable importance from ADM Datamart data.}
\usage{
admVarImp(dmModels, dmPredictors, facets = NULL)
}
\arguments{
\item{dmModels}{Data frame (or anything else that converts to
\code{data.table}) with the ADM datamart model data. Only used
when there are facets, so can be NULL.}

\item{dmPredictors}{Data frame (or anything else that converts to
\code{data.table}) with the ADM datamart predictor data.}

\item{facets}{Optional list of names from the model data to aggregate
up to.}
}
\value{
A \code{data.table}) with \code{PredictorName}, \code{Importance}
and \code{Rank} plus columns for each of the \code{facets} if supplied.
}
\description{
Uses the predictor binning from the ADM Datamart data to find the
relative importance. The importance is given by the distance to the
average log odds, per predictor. Next to importance it also returns
the univariate predictor performance (AUC).
}
\examples{
\dontrun{
  models <- readADMDatamartModelExport("~/Downloads")
  preds <- readADMDatamartPredictorExport("~/Downloads", noBinning = F)

  varimp <- admVarImp(models, preds)
}
}
