% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adm.R
\name{admVarImpPlot}
\alias{admVarImpPlot}
\title{A plot from the ADM variable importance data.}
\usage{
admVarImpPlot(varimp)
}
\arguments{
\item{varimp}{\code{data.table} with the variable importance.}
}
\value{
A \code{ggplot} object that can directly be printed.
}
\description{
Creates a \code{ggplot} object from the provided ADM variable importance
\code{data.table} typically created by \code{admVarImp}.
}
\examples{
\dontrun{
  models <- readADMDatamartModelExport("~/Downloads")
  preds <- readADMDatamartPredictorExport("~/Downloads", noBinning = F)

  varimp <- admVarImp(models, preds)
  print (admVarImpPlot(varimp))
}
}
