% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdh_utils.R
\name{auc_from_bincounts}
\alias{auc_from_bincounts}
\title{Calculates AUC from counts of positives and negatives directly.}
\usage{
auc_from_bincounts(pos, neg)
}
\arguments{
\item{pos}{Vector with counts of the positive responses}

\item{neg}{Vector with counts of the negative responses}
}
\value{
The AUC as a value between 0.5 and 1, return 0.5 if there are any issues with the data.
}
\description{
This is an efficient calculation of the AUC directly from an array of positives and negatives. It makes
sure to always return a value between 0.5 and 1.0 and will return 0.5 in case of any issues.
}
\examples{
auc_from_bincounts( c(3,1,0), c(2,0,1))
}
