% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdh_utils.R
\name{aucpr_from_bincounts}
\alias{aucpr_from_bincounts}
\title{Calculates PR AUC (precision-recall) from counts of positives and negatives directly.}
\usage{
aucpr_from_bincounts(pos, neg, probs = pos/(pos + neg))
}
\arguments{
\item{pos}{Vector with counts of the positive responses}

\item{neg}{Vector with counts of the negative responses}

\item{probs}{Optional vector of probabilities, defaults to pos/(pos+neg). Used to order the response bins.}
}
\value{
The PR AUC as a value between 0.5 and 1.
}
\description{
This is an efficient calculation of the area under the PR curve directly from an array of positives and negatives. It makes
sure to always return a value between 0.5 and 1.0 and will return 0.5 when there is just one
groundtruth label.
}
\examples{
auc_from_bincounts( c(3,1,0), c(2,0,1))
}
