% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdh_utils.R
\name{aucpr_from_probs}
\alias{aucpr_from_probs}
\title{Calculates PR AUC (precision-recall) from an array of truth values and predictions.}
\usage{
aucpr_from_probs(groundtruth, probs)
}
\arguments{
\item{groundtruth}{The 'true' values, with just 2 different values}

\item{probs}{The predictions, as a numeric vector as the same length as \code{groundtruth}}
}
\value{
The AUC as a value between 0.0 and 1.0
}
\description{
Calculates the area under the PR curve from an array of truth values and predictions.
Returns 0.5 when there is just one groundtruth label.
}
\examples{
aucpr_from_probs( c("yes", "yes", "no"), c(0.6, 0.2, 0.2))
}
