% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adm.R
\name{defaultPredictorCategorization}
\alias{defaultPredictorCategorization}
\title{Default predictor categorization. Return a category label given a predictor
name.}
\usage{
defaultPredictorCategorization(p, topLevelLabel = "TopLevel")
}
\arguments{
\item{p}{Predictor name}

\item{topLevelLabel}{Label for the items w/o a dot, defaults to "TopLevel"}
}
\value{
Name of the predictor category
}
\description{
Returns predictor category as the text before the first dot, or the
given label for toplevel if there is no dot.
}
\examples{
defaultPredictorCategorization(c("Customer.Age", "ExternalScores.S1", "Income"))
}
