% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adm.R
\name{expandEmbeddedJSONContext}
\alias{expandEmbeddedJSONContext}
\title{Expand embedded JSON in a column of a data.table into separate columns.}
\usage{
expandEmbeddedJSONContext(dt, fieldName = "Name")
}
\arguments{
\item{dt}{The data.table to use}

\item{fieldName}{The field name to look for embedded JSON. Defaults to Name.}
}
\value{
A data.table with the extra columns. If there is no JSON in the
field returns the original data.table directly.
}
\description{
Mostly
used to expand Name + Treatment that are embedded in the Name field as JSON
strings. Other ADM context key modifications will also show as JSON in the
Name fields, this utility will peel it apart.
}
\details{
Called as part of the standard model read, but exposed so it can also be
applied when reading data from different sources like a CSV file.
}
\examples{
\dontrun{models <- expandEmbeddedJSONContext(models)}
}
