% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adm.R
\name{filterActiveOnly}
\alias{filterActiveOnly}
\title{Subset the datamart predictor data to only the active predictors.}
\usage{
filterActiveOnly(dt, reverse = F)
}
\arguments{
\item{dt}{Predictor datamart table}

\item{reverse}{Boolean to flip the operation.}
}
\value{
A \code{data.table}.
}
\description{
Subset the datamart predictor data to only the active predictors.
}
\examples{
data(adm_datamart)
filterActiveOnly(adm_datamart$predictordata)
}
