% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adm.R
\name{filterInactiveOnly}
\alias{filterInactiveOnly}
\title{Subset the datamart predictor data to only the inactive predictors.}
\usage{
filterInactiveOnly(dt, reverse = F)
}
\arguments{
\item{dt}{Predictor datamart table}

\item{reverse}{Boolean to flip the operation.}
}
\value{
A \code{data.table}.
}
\description{
Subset the datamart predictor data to only the inactive predictors.
}
\examples{
data(adm_datamart)
filterInactiveOnly(adm_datamart$predictordata)
}
