% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adm.R
\name{filterLatestSnapshotOnly}
\alias{filterLatestSnapshotOnly}
\title{Subset the provided datamart data to just the latest snapshot per model.}
\usage{
filterLatestSnapshotOnly(dt)
}
\arguments{
\item{dt}{The \code{data.table} with the datamart data.}
}
\value{
A \code{data.table} with just the latest snapshots per model.
}
\description{
If there is just one snapshot, nothing will change. It works for both
model and predictor tables. If there is no snapshottime field, it will
not do anything.
}
\examples{
data(adm_datamart)
filterLatestSnapshotOnly(adm_datamart$modeldata)
}
