% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adm.R
\name{filterPredictorBinning}
\alias{filterPredictorBinning}
\title{Remove the binning information from the provided datamart predictor data so
it only is predictor level, not more detailed. This is commonly used in
plotting functions.}
\usage{
filterPredictorBinning(dt)
}
\arguments{
\item{dt}{The \code{data.table} with the datamart predictor data.}
}
\value{
A \code{data.table} without the predictor binning.
}
\description{
Remove the binning information from the provided datamart predictor data so
it only is predictor level, not more detailed. This is commonly used in
plotting functions.
}
\examples{
data(adm_datamart)
filterPredictorBinning(adm_datamart$predictordata)
}
