% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adm.R
\name{getActiveRanges}
\alias{getActiveRanges}
\title{Get the active (reachable) score range of one or more models.}
\usage{
getActiveRanges(dm)
}
\arguments{
\item{dm}{An ADM Datamart object (list with models and predictors)}
}
\value{
A list with the min and max score, and the min and max bin index for
each of the models. Apply \code{rbindlist} to the result to get as a table.
}
\description{
Returns a model ID indexed list with the min/max scores as well as the
lower/upper bin indices of the classifiers of the models provided via the
datamart passed in.
Note this method does not support multiple snapshots in the predictor binning
although this could easily be addressed by keying on both ModelID + SnapshotTime
but this situation is very uncommon.
}
\examples{
\dontrun{
  dm <- ADMDatamart("~/Downloads")
  activeRanges <- getActiveRanges(dm)
}
}
