% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportUtilsJSON.R
\name{getJSONModelContextAsString}
\alias{getJSONModelContextAsString}
\title{Converts a model partition (context) from a JSON representation to a string}
\usage{
getJSONModelContextAsString(p)
}
\arguments{
\item{p}{Model partition as a JSON formatted string}
}
\value{
String representation of the context
}
\description{
Converts a model partition (context) from a JSON representation to a string
}
\examples{
library(jsonlite); getJSONModelContextAsString( "{\\"partition\\":{\\"Color\\":\\"Green\\", \\"Size\\":\\"120\\"}}")
}
