% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdh_utils.R
\name{getModelPerformanceOverview}
\alias{getModelPerformanceOverview}
\title{Helper function to return the actual and reported performance of ADM models.}
\usage{
getModelPerformanceOverview(
  dmModels = NULL,
  dmPredictors = NULL,
  jsonPartitions = NULL
)
}
\arguments{
\item{dmModels}{A \code{data.table} with (possibly a subset of) the models exported from the ADM Datamart (\code{Data-Decision-ADM-ModelSnapshot}).}

\item{dmPredictors}{A \code{data.table} with the predictors exported from the ADM Datamart (\code{Data-Decision-ADM-PredictorBinningSnapshot}).}

\item{jsonPartitions}{A list with the (possibly subset) partitions data from the ADM Factory table}
}
\value{
A \code{data.table} with model name, stored performance, actual performance, number of responses and the score range.
}
\description{
In the actual performance the actual score range of the model is
taken into account, while for the stored (reported) performance that is not always the case. This function is mainly in support of validating
a fix to the reported performance.
}
