% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportUtilsDatamart.R
\name{getPredictorsForModelsFromDatamart}
\alias{getPredictorsForModelsFromDatamart}
\title{Retrieves predictor data from the ADM Datamart.}
\usage{
getPredictorsForModelsFromDatamart(conn, models = NULL,
  mostRecentOnly = T, verbose = T)
}
\arguments{
\item{conn}{Connection to the database}

\item{models}{List of models as a \code{data.table} typically obtained from a call to \code{getModelsFromDatamart}.}

\item{mostRecentOnly}{Only return results for latest snapshot. Currently this works as a global filter, which is ok
because the snapshot agent will also snapshot all models wholesale. If and when that becomes more granular, we may
need to make this more subtle, so take the latest after grouping by model ID.}

\item{verbose}{Set to \code{TRUE} to show database queries.}
}
\value{
A \code{data.table} with the ADM predictor details. The names are lowercased so processing of the results is not dependent on DB type.
}
\description{
Typically, predictor data for a certain set of models obtained from \code{getModelsFromDatamart}. It is possible to
retrieve all historical data but by default it only retrieves the most recent snapshot data.
}
\examples{
\dontrun{models <- getModelsFromDatamart(conn); preds <- getPredictorsForModelsFromDatamart(conn, models)}
}
