% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportUtilsJSON.R
\name{getScorecardFromSnapshot}
\alias{getScorecardFromSnapshot}
\title{Convert a model snapshot into a human-readable as well as machine-executable
PMML Scorecard.}
\usage{
getScorecardFromSnapshot(modelSnapshot, name = "")
}
\arguments{
\item{modelSnapshot}{The encoded model snapshot string from the model table.}

\item{name}{Optional name to be used in the generated PMML. Will be used in
the header of the PMML.}
}
\value{
A 4-element \code{list} with the human-readable scorecard in
  $scorecard as a \code{data.table}, a human-readable mapping of the score to
  a propensity in $mapping, also as a \code{data.table}, an executable PMML
  in $pmml in the form of a string and additional the full internal binning.
}
\description{
This works off the model snapshot data stored in the \code{pymodeldata} in
the ADM model table.
}
\details{
This function really is just a convencience wrapper around
\code{getScoringModelFromJSONFactoryString}. It first decodes the data
then delegates to that function.
}
