% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hds}
\alias{hds}
\title{A sample export of a Historical Dataset to illustrate how to import and analyze this.}
\format{
A \code{data.table} with 12062 decisions with associated outcomes.
}
\source{
CDHSample
}
\usage{
hds
}
\description{
A \code{data.table} containing a sample of HDS data.
}
\keyword{datasets}
