% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdh_utils.R
\name{lift}
\alias{lift}
\title{Calculates lift from counts of positives and negatives.}
\usage{
lift(pos, neg)
}
\arguments{
\item{pos}{Vector with counts of the positive responses}

\item{neg}{Vector with counts of the negative responses}
}
\value{
A vector with lift values. Lift can be any positive number. A return value of 2 means 200% lift.
}
\description{
Lift is defined as the success rate in a single bin divided by the overall success rate.
}
\examples{
100*lift(c(0,119,59,69,0), c(50,387,105,40,37))
}
