% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotADMBinning}
\alias{plotADMBinning}
\title{Create binning plot for any predictor.}
\usage{
plotADMBinning(binning, useSmartLabels = T)
}
\arguments{
\item{binning}{Binning of the predictor.}

\item{useSmartLabels}{If true (default) it will abbreviate lengthy symbols
and for numerics it will re-create a shorter interval notation from the
lower and upper bounds. This will generally look better than the labels in
the binning. Turn it off to use the default bin labels. Smart labels do not work
well with additional faceting, so turn it off in that case as well.}
}
\value{
A \code{ggplot} object that can either be printed directly, or to
which additional decoration (e.g. coloring or labels) can be added first.
}
\description{
Combined bar-line plot with bars indicating volume and lines propensity.
}
\examples{
\dontrun{
plotADMBinning(predictorbinning[PredictorName=="NetWealth"])
}
}
