% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotADMModelPerformanceOverTime}
\alias{plotADMModelPerformanceOverTime}
\title{Plot ADM Model Performance over time}
\usage{
plotADMModelPerformanceOverTime(
  modeldata,
  aggregation = intersect(c("Issue", "Group", "Name", "Treatment"), names(modeldata)),
  facets = c("ConfigurationName")
)
}
\arguments{
\item{modeldata}{ADM Datamart model data.}

\item{aggregation}{Vector of field names to aggregate by. There will be a
trend line for each combination of these fields. Defaults to
"Issue","Group","Name","Treatment" (if present).}

\item{facets}{Optional vector of additional fields for faceting. Defaults to
"ConfigurationName"}
}
\value{
A \code{ggplot} object that can either be printed directly, or to
which additional decoration (e.g. coloring or labels) can be added first.
}
\description{
Plot ADM Model Performance over time
}
\examples{
\dontrun{
plotADMModelSuccessRateOverTime(mdls)
}
}
