% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotADMPerformanceSuccessRateBoxPlot}
\alias{plotADMPerformanceSuccessRateBoxPlot}
\title{Create box plot of performance x success rate from ADM datamart model data.}
\usage{
plotADMPerformanceSuccessRateBoxPlot(
  modeldata,
  primaryCriterion = "ConfigurationName",
  facets = c()
)
}
\arguments{
\item{modeldata}{ADM Datamart model data.}

\item{primaryCriterion}{Field to separate by; there will be separate box plots
for each of the values of this field. Can only be a single field. Defaults
to "ConfigurationName".}

\item{facets}{Optional vector of additional fields for faceting. Defaults to empty.}
}
\value{
A \code{ggplot} object that can either be printed directly, or to
which additional decoration (e.g. coloring or labels) can be added first.
}
\description{
Create box plot of performance x success rate from ADM datamart model data.
}
\examples{
\dontrun{
plotADMPerformanceSuccessRateBoxPlot(admdatamart_models, "ConfigurationName") +
  scale_x_continuous(limits = c(50, 70), name = "Performance") + coord_flip()
}
}
