% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotADMPerformanceSuccessRateBubbleChart}
\alias{plotADMPerformanceSuccessRateBubbleChart}
\title{Create bubble chart from ADM datamart model data.}
\usage{
plotADMPerformanceSuccessRateBubbleChart(
  modeldata,
  aggregation = intersect(c("Issue", "Group", "Name", "Treatment"), names(modeldata)),
  facets = c()
)
}
\arguments{
\item{modeldata}{ADM Datamart model data.}

\item{aggregation}{Vector of field names to aggregate by. Defaults to
"Issue","Group","Name","Treatment" (if present).}

\item{facets}{Optional vector of fields for faceting. Defaults to empty.}
}
\value{
A \code{ggplot} object that can either be printed directly, or to
which additional decoration (e.g. coloring or labels) can be added first.
}
\description{
Creates \code{ggplot} scatterplot from success rate x performance of the
models.
}
\examples{
\dontrun{
plotADMPerformanceSuccessRateBubbleChart(mdls, facets = "ConfigurationName")
}
}
