% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotADMPredictorImportance}
\alias{plotADMPredictorImportance}
\title{Create boxplots per predictor with performance range in models.}
\usage{
plotADMPredictorImportance(
  predictordata,
  modeldata = NULL,
  facets = ifelse(is.null(modeldata), "", "ConfigurationName"),
  limit = .Machine$integer.max,
  predictorClassifier = defaultPredictorCategorization,
  categoryAggregateView = F,
  maxNameLength = .Machine$integer.max,
  showAsBoxPlot = T
)
}
\arguments{
\item{predictordata}{ADM Datamart predictor data. Binning data not used and
can be left out.}

\item{modeldata}{Optional ADM Datamart model data, necessary when using
model facets like Channel, ConfigurationName etc.}

\item{facets}{Optional vector of additional fields for faceting. Defaults to
"ConfigurationName" if model data present, otherwise empty.}

\item{limit}{Optional limit of number of predictors. Useful if there are
many predictors to keep the plot readable. Use a negative value to swap the
order, so showing the least performing predictors first.}

\item{predictorClassifier}{Optional function that returns the predictor
categories from the names. Defaults to a function that returns the text
before the first dot.}

\item{categoryAggregateView}{If TRUE shows an overview by predictor categories
instead of the individual predictors.}

\item{maxNameLength}{Max length of predictor names. Above this length
they will be abbreviated.}

\item{showAsBoxPlot}{When TRUE (default), will show as a box plot and sort
by median. When FALSE will show as a simple bar chart and use the weighted
mean (weighted by response count, which is more accurate than the box plot
view).}
}
\value{
A \code{ggplot} object that can either be printed directly, or to
which additional decoration (e.g. coloring or labels) can be added first.
}
\description{
An additional diamond is added to the boxplots to indicate weighted
performance (by response counts in the models).
}
\examples{
\dontrun{
plotADMPredictorImportance(modelPredictorBins, limit = 40) + xlab("AUC")
}
}
