% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotADMPredictorPerformance}
\alias{plotADMPredictorPerformance}
\title{Create boxplots per predictor with performance range in models.}
\usage{
plotADMPredictorPerformance(
  predictordata,
  modeldata = NULL,
  facets = ifelse(is.null(modeldata), "", "ConfigurationName"),
  limit = .Machine$integer.max,
  predictorClassifier = defaultPredictorCategorization,
  categoryAggregateView = F,
  predictorTypeClassifier = ifelse(categoryAggregateView, F, identity),
  showWeightedPerformance = FALSE
)
}
\arguments{
\item{predictordata}{ADM Datamart predictor data. Binning data not used and
can be left out.}

\item{modeldata}{Optional ADM Datamart model data, necessary when using
model facets like Channel, ConfigurationName etc.}

\item{facets}{Optional vector of additional fields for faceting. Defaults to
"ConfigurationName" if model data present, otherwise empty.}

\item{limit}{Optional limit of number of predictors. Useful if there are
many predictors to keep the plot readable. Use a negative value to swap the
order, so showing the least performing predictors first.}

\item{predictorClassifier}{Optional function that returns the predictor
categories from the names. Defaults to a function that returns the text
before the first dot.}

\item{categoryAggregateView}{If TRUE shows an overview by predictor categories
instead of the individual predictors.}

\item{predictorTypeClassifier}{Optional function that returns the type
of the predictor. Defaults to the type in the data, i.e. "symbolic" or
"numeric". This is used to set the line color and shading of the boxplots.
If FALSE the predictor type is not used for the aesthetics of the
boxplot. Defaults to the identity function if not viewed by category.}

\item{showWeightedPerformance}{Annotates the boxplots with a little
indicator that shows the weighted average performance of the predictor.
Defaults to FALSE.}
}
\value{
A \code{ggplot} object that can either be printed directly, or to
which additional decoration (e.g. coloring or labels) can be added first.
}
\description{
An additional diamond is added to the boxplots to indicate weighted
performance (by response counts in the models).
}
\examples{
\dontrun{
plotADMPredictorPerformance(modelPredictorBins, limit = 40) + xlab("AUC")
}
}
