% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotCumulativeLift}
\alias{plotCumulativeLift}
\title{Create cum lift plot for the model classifier.}
\usage{
plotCumulativeLift(binning)
}
\arguments{
\item{binning}{Binning of the classifier (technically it could
be any predictor).}
}
\value{
A \code{ggplot} object that can either be printed directly, or to
which additional decoration (e.g. coloring or labels) can be added first.
}
\description{
Create cum lift plot for the model classifier.
}
\examples{
\dontrun{
plotCumulativeLift(classifierBinning)
}
}
