% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotPerformanceSuccessRateBoxPlot}
\alias{plotPerformanceSuccessRateBoxPlot}
\title{Create box plot of performance x success rate from ADM datamart model data.}
\usage{
plotPerformanceSuccessRateBoxPlot(
  datamart,
  primaryCriterion = "ConfigurationName",
  facets = c(),
  filter = identity
)
}
\arguments{
\item{datamart}{ADM Datamart model data.}

\item{primaryCriterion}{Field to separate by; there will be separate box plots
for each of the values of this field. Can only be a single field. Defaults
to "ConfigurationName".}

\item{facets}{Optional vector of additional fields for faceting. Defaults to empty.}

\item{filter}{Optional filter to apply on the model data}
}
\value{
A \code{ggplot} object that can either be printed directly, or to
which additional decoration (e.g. coloring or labels) can be added first.
}
\description{
Create box plot of performance x success rate from ADM datamart model data.
}
\examples{
\dontrun{
plotPerformanceSuccessRateBoxPlot(adm_datamart_models, "ConfigurationName") +
  scale_x_continuous(limits = c(50, 70), name = "Performance") + coord_flip()
}
}
