% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotPerformanceSuccessRateBubbleChart}
\alias{plotPerformanceSuccessRateBubbleChart}
\title{Create bubble chart from ADM datamart model data.}
\usage{
plotPerformanceSuccessRateBubbleChart(
  datamart,
  aggregation = intersect(c("Issue", "Group", "Name", "Treatment"),
    names(datamart$modeldata)),
  facets = c(),
  color = "Name",
  filter = identity
)
}
\arguments{
\item{datamart}{ADM Datamart model data.}

\item{aggregation}{Vector of field names to aggregate by. Defaults to
"Issue","Group","Name","Treatment" (if present).}

\item{facets}{Optional vector of fields for faceting. Defaults to empty.}

\item{color}{Name of field to use to color the bubbles. Defaults to "Name"
but you could use for example "Group" or "Issue" instead, or use some
custom field that indicates something else alltogether.}

\item{filter}{Optional filter to apply on the model data}
}
\value{
A \code{ggplot} object that can either be printed directly, or to
which additional decoration (e.g. coloring or labels) can be added first.
}
\description{
Creates \code{ggplot} scatterplot from success rate x performance of the
models.
}
\examples{
\dontrun{
plotPerformanceSuccessRateBubbleChart(mdls, facets = "ConfigurationName")
}
}
