% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotPredictorImportance}
\alias{plotPredictorImportance}
\title{Create boxplots per predictor with performance range in models.}
\usage{
plotPredictorImportance(
  datamart,
  facets = ifelse(is.null(datamart$modeldata), "", "ConfigurationName"),
  limit = .Machine$integer.max,
  categoryAggregateView = F,
  maxNameLength = .Machine$integer.max,
  showAsBoxPlot = T,
  activeOnly = (limit > 0),
  filter = identity
)
}
\arguments{
\item{datamart}{ADM Datamart data.}

\item{facets}{Optional vector of additional fields for faceting. Defaults to
"ConfigurationName" if model data present, otherwise empty.}

\item{limit}{Optional limit of number of predictors. Useful if there are
many predictors to keep the plot readable. Use a negative value to swap the
order, so showing the least performing predictors first.}

\item{categoryAggregateView}{If TRUE shows an overview by predictor categories
instead of the individual predictors.}

\item{maxNameLength}{Max length of predictor names. Above this length
they will be abbreviated.}

\item{showAsBoxPlot}{When TRUE (default), will show as a box plot and sort
by median. When FALSE will show as a simple bar chart and use the weighted
mean (weighted by response count, which is more accurate than the box plot
view).}

\item{activeOnly}{When TRUE, only considering predictors that are active. When
FALSE, using all predictors. Defaults to TRUE when the limit is positive, to
FALSE when it is negative.}

\item{filter}{Optional filter to apply on the model data}
}
\value{
A \code{ggplot} object that can either be printed directly, or to
which additional decoration (e.g. coloring or labels) can be added first.
}
\description{
An additional diamond is added to the boxplots to indicate weighted
performance (by response counts in the models).
}
\examples{
\dontrun{
plotPredictorImportance(modelPredictorBins, limit = 40) + xlab("AUC")
}
}
