% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotSuccessRateOverTime}
\alias{plotSuccessRateOverTime}
\title{Plot ADM Model Success Rate over time}
\usage{
plotSuccessRateOverTime(
  datamart,
  aggregation = intersect(c("Issue", "Group", "Name", "Treatment"),
    names(datamart$modeldata)),
  facets = c("ConfigurationName"),
  filter = identity
)
}
\arguments{
\item{datamart}{ADM Datamart model data.}

\item{aggregation}{Vector of field names to aggregate by. There will be a
trend line for each combination of these fields. Defaults to
"Issue","Group","Name","Treatment" (if present).}

\item{facets}{Optional vector of additional fields for faceting. Defaults to
"ConfigurationName"}

\item{filter}{Optional filter to apply on the model data}
}
\value{
A \code{ggplot} object that can either be printed directly, or to
which additional decoration (e.g. coloring or labels) can be added first.
}
\description{
Plot ADM Model Success Rate over time
}
\examples{
\dontrun{
plotSuccessRateOverTime(mdls)
}
}
