% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{printADMPredictorInfo}
\alias{printADMPredictorInfo}
\title{HTML snippet with predictor info.}
\usage{
printADMPredictorInfo(predictorName, binning, extra = list())
}
\arguments{
\item{predictorName}{Name of the predictor.}

\item{binning}{Binning of the predictor.}

\item{extra}{Optional list with extra key-value pairs added to the list.}
}
\description{
Prints (cat) HTML snippet with predictor level info from a \code{data.table} with
the predictor binning.
}
