% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdh_utils.R
\name{readADMDatamartPredictorExport}
\alias{readADMDatamartPredictorExport}
\title{Read export of ADM predictor data.}
\usage{
readADMDatamartPredictorExport(
  srcFolder = ".",
  instancename = "Data-Decision-ADM-PredictorBinningSnapshot_PredictorBinningSnapshot",
  noBinning = T,
  latestOnly = T,
  tmpFolder = tempdir(check = T)
)
}
\arguments{
\item{srcFolder}{Optional folder to look for the file (defaults to the
current folder)}

\item{instancename}{Name of the file w/o the timestamp, in Pega format
<Applies To>_<Instance Name>, or the complete filename including timestamp
and zip extension as exported from Pega. Defaults to the Pega generated
name of the dataset: \code{Data-Decision-ADM-PredictorBinningSnapshot_PredictorBinningSnapshot}.}

\item{noBinning}{If TRUE (the default), skip reading the predictor binning data and just
return predictor level data. For a detailed view with all binning data, set
to FALSE.}

\item{latestOnly}{If TRUE (the default) only the most recent snapshot for every model
is read. To return all data over time (if available), set to FALSE.}

\item{tmpFolder}{Optional folder to store the unzipped data (defaults to a
temp folder)}
}
\value{
A \code{data.table} with the ADM model data
}
\description{
This is a specialized version of \code{readDSExport}
that defaults the dataset name, leaves out internal fields and by default omits the
predictor binning data. In addition it converts date fields and makes sure
the returned fields have the correct type. All string values are returned
as factors.
}
\examples{
\dontrun{readADMDatamartPredictorExport("~/Downloads")}
}
