% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/offlinereports_utils.R
\name{readDatamartFromFile}
\alias{readDatamartFromFile}
\title{Generic datamart reader.}
\usage{
readDatamartFromFile(file)
}
\arguments{
\item{file}{Name of the file. Can be a dataset export, a csv, a zip or
a parquet file.}
}
\value{
A \code{data.table}
}
\description{
Function primarily used by the offline-model report notebooks. If recognized
as such will return model data with all snapshots, or predictor data for only
the latest snapshot.
}
