% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdh_utils.R
\name{readDSExport}
\alias{readDSExport}
\title{Read a Pega dataset export file.}
\usage{
readDSExport(instancename, srcFolder = ".", tmpFolder = srcFolder)
}
\arguments{
\item{instancename}{Name of the file w/o the timestamp, in Pega format <Applies To>_<Instance Name>, or the
complete filename including timestamp and zip extension as exported from Pega.}

\item{srcFolder}{Optional folder to look for the file (defaults to the current folder)}

\item{tmpFolder}{Optional folder to store the unzipped data (defaults to the source folder)}
}
\value{
A \code{data.table} with the contents
}
\description{
\code{readDSExport} reads a dataset export file as exported and downloaded from Pega. The export file
is formatted as a zipped mulit-line JSON file (\url{http://jsonlines.org/}). \code{readDSExport} will find the
most recent file (Pega appends a datetime stamp), unzip it into a temp folder and read the data into a
\code{data.table}.
}
\examples{
\dontrun{readDSExport("Data-Decision-ADM-ModelSnapshot_AllModelSnapshots")}
\dontrun{readDSExport("Data-Decision-ADM-ModelSnapshot_AllModelSnapshots", "~/Downloads")}
\dontrun{readDSExport("Data-Decision-ADM-ModelSnapshot_AllModelSnapshots_20180316T135038_GMT.zip", "~/Downloads")}
}
