% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adm2pmml.R
\name{reasonCodesAboveMin}
\alias{reasonCodesAboveMin}
\title{PMML Scorecard Reason code specification}
\usage{
reasonCodesAboveMin(n = 3)
}
\arguments{
\item{n}{Number of explanations}
}
\value{
A list with data elements used by the underlying implementation
}
\description{
Returns a datastructure to pass to \code{adm2pmml} and \code{createPMML} that
will result in reason codes in the PMML file for points above the minimum value.
}
\examples{
   reasonCodesAboveMin(5)
}
