% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adm2pmml.R
\name{reasonCodesBelowMax}
\alias{reasonCodesBelowMax}
\title{PMML Scorecard Reason code specification}
\usage{
reasonCodesBelowMax(n = 3)
}
\arguments{
\item{n}{Number of explanations}
}
\value{
A list with data elements used by the underlying implementation
}
\description{
Returns a datastructure to pass to \code{adm2pmml} and \code{createPMML} that
will result in reason codes in the PMML file for points below the maximum value.
}
\examples{
   reasonCodesBelowMax(5)
}
