% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdh_utils.R
\name{standardizeFieldCasing}
\alias{standardizeFieldCasing}
\title{Fix field casing}
\usage{
standardizeFieldCasing(dt)
}
\arguments{
\item{dt}{Data table to apply to}
}
\value{
The names of the input are modified by reference
}
\description{
Uniformly sets the field names for data.tables from DB and dataset export,
capitalizing first letter, fixing common endings.
}
\examples{
DT <- data.table::data.table(id = "abc", pyresponsecount = 1234)
standardizeFieldCasing(DT)
}
