% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdh_utils.R
\name{toPRPCDateTime}
\alias{toPRPCDateTime}
\title{Convert to a Pega date-time string.}
\usage{
toPRPCDateTime(x)
}
\arguments{
\item{x}{A vector of valid date-time objects (valid for \code{\link[base:strptime]{strftime}})}
}
\value{
A vector of string representations in the format used by Pega.
}
\description{
Convert to a Pega date-time string.
}
\examples{
toPRPCDateTime(Sys.time())
}
