% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{userFriendlyADMBinning}
\alias{userFriendlyADMBinning}
\title{Creates user friendly table with binning info.}
\usage{
userFriendlyADMBinning(bins)
}
\arguments{
\item{bins}{Binning of the predictor.}
}
\value{
A \code{data.table} table with user friendly binning data.
}
\description{
Creates user friendly table with binning info.
}
