% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdh_utils.R
\name{writeDSExport}
\alias{writeDSExport}
\title{Write table to a file in the format of the dataset export files.}
\usage{
writeDSExport(x, filename, tmpFolder = tempdir())
}
\arguments{
\item{x}{Data table/frame to export.}

\item{filename}{Filename to export to.}

\item{tmpFolder}{Optional folder to store the multi-line JSON file}
}
\description{
\code{writeDSExport} writes a \code{data.table} in the same zipped,
multi-line JSON format (\url{http://jsonlines.org/}) as the Pega datasets.
}
\examples{
\dontrun{writeDSExport(myData, "~/Downloads/mydata.zip")}
}
