% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdh_utils.R
\name{zratio}
\alias{zratio}
\title{Calculates the Z-Ratio for the success rate from counts of positives and negatives.}
\usage{
zratio(pos, neg)
}
\arguments{
\item{pos}{Vector with counts of the positive responses}

\item{neg}{Vector with counts of the negative responses}
}
\value{
A vector with z-ratio values. Z-ratio can be both positive and negative.
}
\description{
Z-Ratio is the difference between the propensity in the bin with the average propensity, expressed in standard deviations from the average.
}
\examples{
zratio(c(0,119,59,69,0), c(50,387,105,40,37))
}
