% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adm.R
\name{ADMDatamart}
\alias{ADMDatamart}
\title{Generic method to read ADM Datamart data into a list structure.}
\usage{
ADMDatamart(
  modeldata = NULL,
  predictordata = NULL,
  folder = NULL,
  keepSerializedModelData = FALSE
)
}
\arguments{
\item{modeldata}{Location, reference or the actual model table from the
ADM datamart.}

\item{predictordata}{Location, reference or the actual predictor binning
table from the ADM datamart.}

\item{folder}{Optional path for the folder in which to look for the model
and predictor data.}

\item{keepSerializedModelData}{By default the serialized model data is left
out from the model data as this typically is large and only needed for
specific use cases.}
}
\value{
A list that wraps the the two \code{data.table} elements plus
a few flags that indicate whether there are multiple snapshots or not.
}
\description{
Method is very flexible in the arguments. It can take a \code{data.table}
for the model data and the predictor data, but the arguments can also be
pointing to files in CSV, zipped CSV, (LD)JSON, parquet or dataset export formats.
}
