% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{admdatamart_models}
\alias{admdatamart_models}
\title{Sample model snapshots from the ADM Data Mart.}
\format{A \code{data.table} with 30 rows and 22 variables:
\describe{
  \item{pyModelID}{Model ID}
  \item{pySnapshotTime}{Snapshot time}
  ...
}}
\source{
A sample Pega application
}
\usage{
admdatamart_models
}
\description{
A \code{data.table} containing two snapshots of 3 models.
}
\keyword{datasets}
