% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adm.R
\name{admVarImp}
\alias{admVarImp}
\title{Calculate variable importance from ADM Datamart data.}
\usage{
admVarImp(dmModels, dmPredictors, facets = NULL)
}
\arguments{
\item{dmModels}{Data frame (or anything else that converts to
\code{data.table}) with the ADM datamart model data.}

\item{dmPredictors}{Data frame (or anything else that converts to
\code{data.table}) with the ADM datamart predictor data.}

\item{facets}{Optional list of names from the model data to split the
predictor importance by.}
}
\value{
A \code{data.table}) with \code{PredictorName}, \code{Importance}
and \code{Rank} plus columns for each of the \code{facets} if supplied. In
addition there is a column \code{ConfigurationNames} with a concatenation
of all the values for \code{ConfigurationName} in the original model data.
}
\description{
Uses the predictor binning from the ADM Datamart data to find the
relative predictor importance. Uses the exact same cumulative log odds
scores and weighs these by the evidence in the bins.
}
\examples{
\dontrun{
  models <- readADMDatamartModelExport("~/Downloads")
  preds <- readADMDatamartPredictorExport("~/Downloads", noBinning = F)

  varimp <- admVarImp(models, preds)
}
}
