% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adm.R
\name{admVarImp}
\alias{admVarImp}
\title{Calculate variable importance from ADM Datamart data.}
\usage{
admVarImp(
  datamart,
  facets = NULL,
  filter = function(x) {     filterClassifierOnly(x, reverse = T) }
)
}
\arguments{
\item{datamart}{Data frame with the ADM datamart data. Model data is only
used when there are facets, so the model part could in principle be NULL.}

\item{facets}{Optional list of names from the model data to aggregate
up to.}

\item{filter}{Optional filter for the predictor data. Defaults to filter
out the classifiers, so operates on both active and inactive predictors.
Pass in \code{filterActiveOnly} to only consider the active predictors.}
}
\value{
A \code{data.table}) with \code{PredictorName}, \code{Importance}
and \code{Rank} plus columns for each of the \code{facets} if supplied.
}
\description{
Uses the predictor binning from the ADM Datamart data to find the
relative importance. The importance is given by the distance to the
average log odds, per predictor. Next to importance it also returns
the univariate predictor performance (AUC).
}
\examples{
\dontrun{
  dm <- ADMDatamart("~/Downloads")

  varimp <- admVarImp(dm)
}
}
