% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdh_utils.R
\name{auc_from_probs}
\alias{auc_from_probs}
\title{Calculates AUC from an array of truth values and predictions.}
\usage{
auc_from_probs(groundtruth, probs)
}
\arguments{
\item{groundtruth}{The 'true' values, with just 2 different values, passed on to the pROC::roc function.}

\item{probs}{The predictions, as a numeric vector as the same length as \code{groundtruth}}
}
\value{
The AUC as a value between 0.5 and 1, return 0.5 if there are any issues with the data.
}
\description{
Calculates the area under the ROC curve from an array of truth values and predictions, making sure to
always return a value between 0.5 and 1.0 and will return 0.5 in case of any issues.
}
\examples{
auc_from_probs( c("yes", "yes", "no"), c(0.6, 0.2, 0.2))
}
