% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdh_utils.R
\name{fromPRPCDateTime}
\alias{fromPRPCDateTime}
\title{Convert from a Pega date-time string.}
\usage{
fromPRPCDateTime(x)
}
\arguments{
\item{x}{A vector of Pega date-time strings}
}
\value{
An object of class \linkS4class{POSIXlt}
}
\description{
Convert from a Pega date-time string.
}
\examples{
fromPRPCDateTime("20180316T134127.847 GMT")
}
