% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportUtilsJSON.R
\name{getScoringModelFromJSONFactoryString}
\alias{getScoringModelFromJSONFactoryString}
\title{Build a scoring model from the "analyzedData" JSON string that is part of the JSON Factory.}
\usage{
getScoringModelFromJSONFactoryString(analyzedData)
}
\arguments{
\item{analyzedData}{The JSON string with the analyzed (scoring) portion of the ADM Factory JSON.}
}
\value{
A \code{data.table} with the binning.
}
\description{
This part of the ADM Factory string is the only piece that is necessary for scoring. The
scoring model itself is currently just a binning table in internal format. We should see
if this needs to be augmented to make it more of a score card.
}
