% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ihsampledata}
\alias{ihsampledata}
\title{A small sample of IH data to illustrate how to import and analyze this.}
\format{A \code{data.table} with 174 interactions for in total 10 different customers.}
\source{
DMSample
}
\usage{
ihsampledata
}
\description{
A \code{data.table} containing a sample of IH data.
}
\keyword{datasets}
