% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotADMBinning}
\alias{plotADMBinning}
\title{Create binning plot for any predictor.}
\usage{
plotADMBinning(binning)
}
\arguments{
\item{binning}{Binning of the predictor.}
}
\value{
A \code{ggplot} object that can either be printed directly, or to
which additional decoration (e.g. coloring or labels) can be added first.
}
\description{
Combined bar-line plot with bars indicating volume and lines propensity.
}
\examples{
\dontrun{
plotADMBinning(predictorbinning[PredictorName=="NetWealth"])
}
}
