% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotADMPredictorPerformance}
\alias{plotADMPredictorPerformance}
\title{Create boxplots per predictor with performance range in models.}
\usage{
plotADMPredictorPerformance(
  predictordata,
  modeldata = NULL,
  facets = ifelse(is.null(modeldata), "", "ConfigurationName"),
  limit = .Machine$integer.max,
  predictorClassifier = defaultPredictorCategorization
)
}
\arguments{
\item{predictordata}{ADM Datamart predictor data. Binning data not used and
can be left out.}

\item{modeldata}{Optional ADM Datamart model data, necessary when using
model facets like Channel, ConfigurationName etc.}

\item{facets}{Optional vector of additional fields for faceting. Defaults to
"ConfigurationName" if model data present, otherwise empty.}

\item{limit}{Optional limit of number of predictors. Useful if there are
many predictors to keep the plot readable.}

\item{predictorClassifier}{Optional function that returns the predictor
categories from the names. Defaults to a function that returns the text
before the first dot.}
}
\value{
A \code{ggplot} object that can either be printed directly, or to
which additional decoration (e.g. coloring or labels) can be added first.
}
\description{
An additional diamond is added to the boxplots to indicate weighted
performance (by response counts in the models).
}
\examples{
\dontrun{
plotADMPredictorPerformance(modelPredictorBins, limit = 40) + xlab("AUC")
}
}
