% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdh_utils.R
\name{readDSExport}
\alias{readDSExport}
\title{Read a Pega dataset export file.}
\usage{
readDSExport(
  instancename,
  srcFolder = ".",
  tmpFolder = tempdir(check = T),
  excludeComplexTypes = T,
  acceptJSONLines = NULL,
  stringsAsFactors = F,
  verbose = F
)
}
\arguments{
\item{instancename}{Name of the file w/o the timestamp, in Pega format
<Applies To>_<Instance Name>, or the complete filename of the dataset
export file, or a specific JSON file.}

\item{srcFolder}{Optional folder to look for the file (defaults to the
current folder)}

\item{tmpFolder}{Optional folder to store the unzipped data (defaults to a temp folder)}

\item{excludeComplexTypes}{Flag to not return complex embedded types,
defaults to T so not including nested lists/data frames}

\item{acceptJSONLines}{Optional filter function. If given, this is applied
to the lines with JSON data and only those lines for which the function
returns TRUE will be parsed. This is just for efficiency when reading
really big files so you can filter rows early.}

\item{stringsAsFactors}{Logical (default is FALSE). Convert all character columns to factors?}

\item{verbose}{Flag for verbose logging, defaults to FALSE.}
}
\value{
A \code{data.table} with the contents
}
\description{
\code{readDSExport} reads a dataset export file as exported and downloaded
from Pega datasets. This export file is a zipped multi-line JSON file
(\url{http://jsonlines.org/}). \code{readDSExport} will find the most recent
file (Pega appends a datetime stamp) if given a folder, then unzips it into a
temp folder and read the data into a \code{data.table}. You can also specify
a full file name to the export or to a JSON file that you unpacked yourself.
}
\examples{
\dontrun{readDSExport("Data-Decision-ADM-ModelSnapshot_All")}
\dontrun{readDSExport("Data-Decision-ADM-ModelSnapshot_All", "~/Downloads")}
\dontrun{readDSExport("Data-Decision-ADM-ModelSnapshot_All_20180316T135038_GMT.zip",
"~/Downloads")}
\dontrun{readDSExport("~/Downloads/Data-Decision-ADM-ModelSnapshot_All_20180316T135038_GMT.zip")}
\dontrun{readDSExport("SampleApp-SR_DecisionResults.json", "data")}
\dontrun{readDSExport("json$", "create_hds/data")}
}
